web/requirements.txt
Flask
docker
bootstrap-flask
WTForms
Flask-WTF
requests
celery[redis]
flower
python-decouple
pymysql